update finish authorization system

+++ app/Http/Controllers/AdminRatingController.php
@@ -7,17 +7,22 @@ use App\Models\RoomCategory;
 use App\Models\Transaction;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class AdminRatingController extends Controller
 {
     public function index()
     {
-        return view('admin.rating.index', [
-            'title' => 'Rating',
-            'transactions' => Transaction::all(),
-            'room_categories' => RoomCategory::all(),
-            'users' => User::all()
-        ]);
+        if (Gate::allows('11_1')) {
+            return view('admin.rating.index', [
+                'title' => 'Rating',
+                'transactions' => Transaction::all(),
+                'room_categories' => RoomCategory::all(),
+                'users' => User::all()
+            ]);
+        } else {
+            return redirect()->route('admin.error-401');
+        }
     }
 
     public function table()
@@ -30,7 +35,11 @@ class AdminRatingController extends Controller
 
     public function delete($id)
     {
-        Rating::destroy($id);
-        return response()->json(['message' => 'Data deleted successfully!']);
+        if (Gate::allows('10_4')) {
+            Rating::destroy($id);
+            return response()->json(['message' => 'Data deleted successfully!']);
+        } else {
+            return redirect()->route('admin.error-401');
+        }
     }
 }
